让 64 位支持32位编译出的字节码 https://blog.csdn.net/you_lan_hai/article/details/77430116

+++ app/src/main/cpp/lua/ldump.c
@@ -37,7 +37,7 @@ typedef struct {
 #define DumpLiteral(s,D)	DumpBlock(s, sizeof(s) - sizeof(char), D)
 
 
-static void DumpBlock (const void *b, size_t size, DumpState *D) {
+static void DumpBlock (const void *b, uint32_t size, DumpState *D) {
   if (D->status == 0 && size > 0) {
     lua_unlock(D->L);
     D->status = (*D->writer)(D->L, b, size, D->data);
@@ -74,7 +74,7 @@ static void DumpString (const TString *s, DumpState *D) {
   if (s == NULL)
     DumpByte(0, D);
   else {
-    size_t size = tsslen(s) + 1;  /* include trailing '\0' */
+    uint32_t size = tsslen(s) + 1;  /* include trailing '\0' */
     const char *str = getstr(s);
     if (size < 0xFF)
       DumpByte(cast_int(size), D);
@@ -187,7 +187,7 @@ static void DumpHeader (DumpState *D) {
   DumpByte(LUAC_FORMAT, D);
   DumpLiteral(LUAC_DATA, D);
   DumpByte(sizeof(int), D);
-  DumpByte(sizeof(size_t), D);
+  DumpByte(sizeof(uint32_t), D);
   DumpByte(sizeof(Instruction), D);
   DumpByte(sizeof(lua_Integer), D);
   DumpByte(sizeof(lua_Number), D);
